+++ types/portal_region.go
@@ -3,7 +3,7 @@ package types
 import "time"
 
 type PortalRegion struct {
-	PortalRegionID   int64     `json:"portalRegionID"`
+	PortalRegionID   int       `json:"portalRegionID"`
 	PortalRegionName string    `json:"portalRegionName"`
 	CreatedAt        time.Time `json:"createdAt"`
 	UpdatedAt        time.Time `json:"updatedAt"`
